Refactor features folders
