+++ src/features/ubication/list/Ubication.test.tsx
@@ -1,7 +1,7 @@
 import { test, expect } from "vitest";
 import { type Ubication as TUbication } from "@prisma/client";
 
-import Ubication from "./Ubication";
+import { Ubication } from "./Ubication";
 import { render, screen } from "~/utils/test/test-utils";
 
 const mockUbication: TUbication = {
