+++ src/features/ubication/list/Ubication.tsx
@@ -3,11 +3,11 @@ import { PlusCircle, ThermometerSnowflake } from "lucide-react";
 import Link from "next/link";
 
 import { api } from "~/utils/api";
-import FoodList from "../../food/components/FoodList";
+import FoodList from "../../food/list/FoodList";
 import { Loader } from "../../common/components";
 import { Button } from "../../common/components/Buttons";
 
-export default function Ubication({ data }: { data: TUbication }) {
+export function Ubication({ data }: { data: TUbication }) {
   const { data: foods, status } = api.food.getFromUbication.useQuery(
     {
       ubicationId: data.id,
