+++ src/pages/space/[id]/index.tsx
@@ -2,14 +2,12 @@ import Link from "next/link";
 import { useRouter } from "next/router";
 import { Container } from "lucide-react";
 
-import UbicationList from "~/features/ubication/components/UbicationList";
+import UbicationList from "~/features/ubication/list/UbicationList";
 import { api } from "~/utils/api";
 import QueryErrorBoundary from "~/features/common/components/Error/QueryErrorBoundary";
 import { Loader } from "~/features/common/components";
 import { PageLayout } from "~/features/common/components/layout";
-import { EmptyUbications } from "~/features/ubication/components/EmptyUbications";
-
-import { ShareSpace } from "~/features/space/components/ShareSpace";
+import { ShareSpace } from "~/features/space/share/ShareSpace";
 
 export default function Space() {
   const router = useRouter();
@@ -46,14 +44,10 @@ export default function Space() {
             >
               Volver al listado de espacios
             </Link>
-            {space.ubications.length > 0 ? (
-              <UbicationList
-                ubications={space?.ubications}
-                spaceId={spaceId?.toString() ?? ""}
-              />
-            ) : (
-              <EmptyUbications spaceId={spaceId?.toString()} />
-            )}
+            <UbicationList
+              ubications={space?.ubications}
+              spaceId={spaceId?.toString() ?? ""}
+            />
           </>
         )}
       </QueryErrorBoundary>
